requirement.txt
anthropic>=0.8.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
openai>=1.3.0